Fix streaming parsing
So my previous fix only fixed the first result in the stream. All of the others are of a different format. However if we ask for the /last value directly all results look the same.

Didn't see this before because the job updating the bitcoin database was down and it wasn't seeing any updates but that is fixed now.